+++ CHANGELOG.md
@@ -5,11 +5,13 @@
 - Use of hgijson from PyPI.
 - Type of exception raised when baton encounters an unknown exception ([#34](https://github.com/wtsi-hgi/python-baton-wrapper/issues/34)).
 
+
 ### Added
 - Ability to recursively change access controls.
 
 ### Removed
 - Redundant helper scripts.
+- `get_all` method from `DataObjectReplicaCollection`.
 
 
 ## 0.5.0 - 2016-05-05